requirements.txt
pandas==2.2.2
openai==1.52.0
python-dotenv==1.0.1
httpx<0.28.0
numpy>=1.26.0
